Art Kit
Kit for aircraft artits and creators to add wing air livery to their airplanes.
W I N G  A I R should be placed just above windows close to the front door.
W N G X X X is aircraft registration number, 727 for Boeing 727, 320 for Airbus 320 and so on.
Vertical Wings for the vertical stabilizer. Horizontal wings can be added to the side of the engines.